requirements.txt
pandas
matplotlib
seaborn
tqdm
imagehash
scikit-image
Pillow
opencv-python
torch
torchvision
pytorch-lightning
